refactor: 당첨된 등수별 로또 개수를 반환하는 DTO record로 변경
