+++ src/main/java/lotto/dto/LottoRankResultDTO.java
@@ -1,0 +1,14 @@
+package lotto.dto;
+
+import lotto.model.LottoRank;
+import lotto.model.LottoRankResult;
+
+public record LottoRankResultDTO(LottoRank rank, LottoRankResult result) {
+    public String getRankMessage() {
+        return rank.getMessage();
+    }
+
+    public int getCount() {
+        return result.getCountByLottoRank(rank);
+    }
+}
