+++ src/main/java/lotto/model/LottoRankResult.java
@@ -1,4 +1,4 @@
-package lotto.dto;
+package lotto.model;
 
 import static java.util.Collections.synchronizedMap;
 
@@ -7,14 +7,11 @@ import java.util.EnumMap;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
-import lotto.model.Lotto;
-import lotto.model.LottoRank;
-import lotto.model.WinningLotto;
 
-public class WinningLottoResult {
+public class LottoRankResult {
     private final Map<LottoRank, Integer> result;
 
-    public WinningLottoResult() {
+    public LottoRankResult() {
         List<LottoRank> ranks = Arrays.asList(LottoRank.values());
         result = synchronizedMap(new EnumMap<>(LottoRank.class));
         ranks.forEach(rank -> result.put(rank, 0));
